requirements.txt
django
djangorestframework
psycopg2-binary
drf-yasg[validation]
opencv-python
pyttsx3
numpy
Pillow
pytesseract
tesseract
gtts
playsound
pipwin
SpeechRecognition
pydub
django-heroku
whitenoise
gunicorn